Add support for getThemeLogo
